refactor: 이미지 업로드 로직을 useImageUpload로 분리
